fixed bug described in the issue
